Moved the watcher code to a new file
